Add files via upload
My Scratch project involves a dog asking for your name and inviting you to get fruit. The dog will walk away and then reappear in a new backdrop. The dog will then instruct you to put the fruit in the bowl, which will transform into a fruit bowl. Unfortunately, the fruit and bowl sometimes do not hide when resetting the game due to bad code. I've tried to fix it, though, with no change. :/